requirements.txt
djangodjango-cors-headersdjangorestframeworkdjangorestframework-simplejwtpillowpsycopg2-binaryrequestsscikit-learntorchtransformers